requirements.txt
imbalanced-learn==0.13.0
joblib==1.4.2
matplotlib-inline==0.1.7
scipy==1.15.2
scikit-learn==1.3.2
xgboost==3.0.2
tqdm==4.67.1
seaborn==0.13.2